Move utils to algorithms
