+++ algebra/u256/src/algorithms/binary_operator_macros.rs
