+++ algebra/u256/src/algorithms/mod.rs
@@ -1,5 +1,7 @@
+mod binary_operator_macros;
 mod knuth_division;
 mod lehmer_gcd;
+pub mod limb_operations;
 
 pub use knuth_division::{divrem_nby1, divrem_nbym};
 pub(crate) use lehmer_gcd::inv_mod;
